refaktorerte metode fra innkalling til felles metadataview
